chore(quality): execute CLEANUP P0+P1 — dead edges, god-file splits, hidden-coupling extractions
  10 dead Cargo edges cut; dead cfg blocks removed; lock audit (23 sites, all
  tests); sqlite.rs and retrieval.rs split under the 1,200 ceiling;
  plan_grammar/synthesis_common/middleware-shared extractions make git-observed
  hidden coupling structural; sovereign-contracts docs 798→0; clippy --fix
  sweep (baseline 2,555→1,637); api snapshots for the contract leaves
  (hub crates parked on a nightly ICE in lance-index). Fixes SYSTEM_OVERVIEW's
  fictional planner-annotation grammar. All gates green.

+++ sovereign/crates/sovereign-core/src/runtime/handlers/generative.rs
@@ -100,7 +100,7 @@ impl Runtime {
         let conversation_id_owned = conversation_id.to_string();
         let message_id_for_persist = message_id.clone();
         let context_window = self.inference.effective_context_size();
-        let max_tokens_budget = request.max_tokens.map(|t| t as usize);
+        let max_tokens_budget = request.max_tokens.map(|t| t);
         let started = std::time::Instant::now();
         let (tx, rx) = tokio::sync::mpsc::unbounded_channel::<Result<String>>();
